Abort unit tests if metarig.py is modified.
After unit tests are run, run `git checkout` to restore metarig.py to
it's pre unit test state.

metarig.py is modified when the unit tests are run. This change aborts
the unit tests if metarig.py is modified. If it is not modified, it
allows the unit tests to run, and then git checkouts metarig.py to
restore it to it's pre-unit test state.