Update documentation to state where ue2rigify stores new templates
# What was changed

The ue2rigify documentation now makes mention of where newly created templates are found

# Why it was changed

The "Create New" button, with "male_mannequin" metarig selected, does not create FK and Deform Bone re-mapping nodes for you. In this case, it is useful to copy the "male_mannequin" template folder (for example) and rename it, so you can make your tweaks on this version instead. This avoids users editing the templates that come ootb.

# Note

The path may be different on Linux or MAC OS, therefore, it may be worth adding a list of locations (Windows: ... Linux: ..., MacOS: ...),  or a more general note about where you may find the templates.